+++ TeamDive_Admin/src/style/membership.scss
@@ -18,6 +18,7 @@ $border-color: #ddd;
         width: 100%;
         border-collapse: collapse;
         margin: auto;
+        
 
         th, td {
             padding: 12px;
